+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 * Create permission "Fees/Fines: All Actions permissions". Ref UIU-673.
 * Implement CRUD Manual Patron Block. Ref UIU-674.
 * Create permission "Patron blocks: All permissions". Ref UIU-728.
+* Request queue link should include only open requests. Fixes UIU-716.
 
 ## [2.17.0](https://github.com/folio-org/ui-users/tree/v2.17.0) (2018-10-5)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.16.0...v2.17.0)